# bar-frontend/src/app/api-client/model/categoryDto.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SubcategoryDto } from './subcategoryDto';

export interface CategoryDto {
    id?: number;
    createdBy?: string;
    lastModifiedBy?: string;
    createdDate?: string;
    lastModifiedDate?: string;
    code?: string;
    name?: string;
    iconUrl?: string;
    subcategories?: Array<SubcategoryDto>;
}