# bar-frontend/src/app/api-client/model/itemTagDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ItemTagDto { 
    label?: string;
    cssClass?: string;
}